Revert "Fix #46 (#56)"
This reverts commit f71f296ed3e6d6cbf5f8c0fd8e8114ebb598294c.
