+++ tools/format.sh
@@ -4,4 +4,3 @@ deno fmt src
 deno fmt test
 deno fmt templates
 deno fmt cli
-deno fmt tools
